Preliminary work to support virtual albums.
Architecture change to make Gallery own all of the Media objects.
The self.medias property in Album is not a list of keys into the
Gallery medias dictionary. Methods in Album that iterate or
reference the Album's medias property now take the key value from
the Album medias list and use it to dereference an actual Media
object from the medias dictionary in Gallery.