Pass object list instead of id list to Album init.
Album media sort assumes the media list in the album contains the
actual media objects. Trying to indirectly reference back to the
master media dictionary in the Gallery makes this really ugly.
Since Media objects are pass by reference, we can have the same
objects referenced in both the Gallery media dict and the Album
media list. So changed the Gallery code to pass the list of Media
objects in the Album to the Album init.